+++ lib/braintree/credit_card_verification_gateway.rb
@@ -47,22 +47,39 @@ module Braintree
     def self._create_signature
       [
         {:credit_card => [
-          :cardholder_name, :cvv, :expiration_date, :expiration_month, :expiration_year,
-          :number, {:billing_address => AddressGateway._shared_signature}
+          {:billing_address => AddressGateway._shared_signature},
+          :cardholder_name,
+          :cvv,
+          :expiration_date,
+          :expiration_month,
+          :expiration_year,
+          :number,
+        ]},
+        {:external_vault => [
+          :previous_network_transaction_id,
+          :status,
         ]},
         :intended_transaction_source,
-        {:options => [:amount, :merchant_account_id, :account_type]},
+        {:options => [
+          :account_type,
+          :amount,
+          :merchant_account_id,
+        ]},
         :payment_method_nonce,
+        {:risk_data => [
+          :customer_browser,
+          :customer_ip,
+        ]},
         :three_d_secure_authentication_id,
         {:three_d_secure_pass_thru => [
-            :eci_flag,
-            :cavv,
-            :xid,
-            :three_d_secure_version,
-            :authentication_response,
-            :directory_response,
-            :cavv_algorithm,
-            :ds_transaction_id,
+          :authentication_response,
+          :cavv,
+          :cavv_algorithm,
+          :directory_response,
+          :ds_transaction_id,
+          :eci_flag,
+          :three_d_secure_version,
+          :xid,
         ]},
       ]
     end