+++ lib/braintree/http.rb
@@ -7,8 +7,8 @@ module Braintree
       @config = config
     end
 
-    def delete(_path, query_params = {})
-      path = _path + _build_query_string(query_params)
+    def delete(path, query_params = {})
+      full_path = path + _build_query_string(query_params)
       response = _http_do Net::HTTP::Delete, path
       if response.code.to_i == 200 || response.code.to_i == 204
         true
@@ -19,8 +19,8 @@ module Braintree
       end
     end
 
-    def get(_path, query_params = {})
-      path = _path + _build_query_string(query_params)
+    def get(path, query_params = {})
+      full_path = path + _build_query_string(query_params)
       response = _http_do Net::HTTP::Get, path
       if response.code.to_i == 200 || response.code.to_i == 422
         Xml.hash_from_xml(_body(response))