+++ spec/unit/braintree/base_module_spec.rb
@@ -8,11 +8,11 @@ describe Braintree::BaseModule do
         klass = Class.new { include Braintree::BaseModule }
         klass.return_object_or_raise(:obj) { result }
       rescue Braintree::ValidationsFailed => ex
-      end
       expect(ex).not_to eq(nil)
       expect(ex.error_result).to eq(result)
       expect(ex.inspect).to include(result.inspect)
       expect(ex.to_s).to include(result.inspect)
+      end
     end
   end
 end